Update contracts.md.
Remove mention of Aggregate Signatures. Formally, an aggregate signature scheme would require an aggregated signature on messages \(m_1, .... m_n\) to be of the same size as a signature on a single message. We used some slack in [FOS19] to bring Schnorr and BLS under the same framework but Schnorr/MuSig are definitely not know for its aggregation capabilities. MuSig is a multisignature, which is very different.